+++ .env.example
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_NEYNAR_API_KEY=""
 NEXT_PUBLIC_AMPLITUDE_API_KEY=""
+NEXT_PUBLIC_CLIENT_ID=""
