+++ package.json
@@ -14,7 +14,7 @@
     "@amplitude/analytics-browser": "^2.9.3",
     "@farcaster/hub-nodejs": "^0.11.11",
     "@neynar/nodejs-sdk": "^1.27.0",
-    "@neynar/react": "^0.6.1",
+    "@neynar/react": "^0.6.4",
     "@pigment-css/react": "^0.0.9",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-collapsible": "^1.0.3",
