+++ src/components/search.tsx
@@ -1,13 +1,14 @@
 'use client';
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next13-progressbar';
-import { usePathname } from 'next/navigation';
+import { usePathname, useSearchParams } from 'next/navigation';
 import { SearchIcon, X } from 'lucide-react';
 import * as amplitude from '@amplitude/analytics-browser';
 
 export default function Search() {
   const router = useRouter();
   const path = usePathname();
+  const searchParams = useSearchParams();
   const [identifier, setIdentifier] = useState<string>(
     decodeURIComponent(path.slice(1))
   );
@@ -21,16 +22,23 @@ export default function Search() {
     setIdentifier('');
   };
 
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    amplitude.track('Search made', {
+      identifier,
+      ...Object.fromEntries(searchParams.entries()),
+    });
+
+    const currentParams = new URLSearchParams(searchParams.toString());
+    const url = `/${encodeURIComponent(identifier)}${currentParams.toString() ? `?${currentParams.toString()}` : ''}`;
+
+    router.push(url);
+  };
+
   return (
     <form
       className="flex flex-row items-start p-0 w-full h-[35px]"
-      onSubmit={(e) => {
-        e.preventDefault();
-        amplitude.track('Search made', {
-          identifier,
-        });
-        router.push(`/${encodeURIComponent(identifier)}`);
-      }}
+      onSubmit={handleSubmit}
     >
       <div className="flex flex-row items-center pl-[7.5px] w-full h-[35px] bg-white border border-white relative">
         <input
